+++ src/motor/motor.h
@@ -3,12 +3,12 @@
  * @file    motor.h
  * @author  Stephen Papierski <stephenpapierski@gmail.com>
  * @date    2015-03-26 03:49:13
- * @edited  2015-03-26 03:51:14
+ * @edited  2015-04-28 16:02:07
  */
 
 #ifndef MOTOR_H
 #define MOTOR_H
 
-int motor_update_pid_A(float torqueRef);
+double motor_update_pid_A(float torqueRef);
 
 #endif //MOTOR_H
